System :sepolicy: Fix bootloop on TWRP restore
sepolicy: Fix bootloop on TWRP restore

When TWRP restores a ROM backup, the root directory now has the wrong
selinux
context "system_file" instead of "rootfs" as it should be for Q
system-as-root.

Attempting to boot the restored ROM results in the following
error/denial
and will bootloop if the ROM is enforcing.

Therefore, added sepolicy/vendor/zygote.te on device tree.
As result, neverallow rule on domain.te must change.

Signed-off-by: ashaykapadiya <kapadiyaashay@gmail.com>
Change-Id: I35babe80e0e5cfcf7ffa480919d83f3b39369acf